# contracts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ChangePasswordRequest": {
        "description": "Запит на зміну пароля.",
        "properties": {
          "currentPassword": {
            "description": "Поточний пароль.",
            "type": "string"
          },
          "newPassword": {
            "description": "Новий пароль.",
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "type": "object"
      },
...
      "CreateRoleRequest": {
        "description": "Запит на створення ролі.",
        "properties": {
          "code": {
            "description": "Код ролі.",
            "type": "string"
          },
          "nameL10n": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Назва мовами каталогу.",
            "type": "object"
          },
          "permissionCodes": {
            "description": "Права, що входять у роль.",
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "code",
          "nameL10n",
          "permissionCodes"
        ],
        "type": "object"
      },
...
      "CreateUserRequest": {
        "description": "Запит на створення користувача.",
        "properties": {
          "displayName": {
            "description": "Ім'я для показу; типово збігається з іменем входу.",
            "type": [
              "null",
              "string"
            ]
          },
          "initialPassword": {
            "description": "Разовий пароль локального запису.",
            "type": [
              "null",
              "string"
            ]
          },
          "provider": {
            "description": "`Windows` або `Local`.",
            "type": "string"
          },
          "roleCodes": {
            "description": "Ролі, які призначити одразу.",
            "items": {
              "type": "string"
            },
            "type": [
              "null",
              "array"
            ]
          },
          "sid": {
            "description": "SID доменного користувача; `null` для локального.",
            "type": [
              "null",
              "string"
            ]
          },
          "userName": {
            "description": "Ім'я входу.",
            "type": "string"
          }
        },
        "required": [
          "userName",
          "provider",
          "sid"
        ],
        "type": "object"
      },
...
      "GrantLevel": {
        "description": "Рівень ресурсного гранта. Порядок значень значущий: більше = ширше.",
        "enum": [
          "None",
          "Read",
          "Write",
          "Submit",
          "Approve",
          "Manage"
        ]
      },
...
      "PagedResultOfUserView": {
        "description": "Сторінка результатів. Ендпоінтів, що повертають «усе», не існує —\r\nперевіряється архітектурним тестом.",
        "properties": {
          "items": {
            "description": "Елементи сторінки.",
            "items": {
              "$ref": "#/components/schemas/UserView"
            },
            "type": "array"
          },
          "nextCursor": {
            "description": "Курсор наступної сторінки; `null` — кінець.",
            "type": [
              "null",
              "string"
            ]
          },
          "totalCount": {
            "description": "Загальна кількість; `null`, якщо підрахунок дорогий.",
            "format": "int32",
            "type": [
              "null",
              "integer"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor",
          "totalCount"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "format": "int32",
            "type": [
              "null",
              "integer"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "type": "object"
      },
...
      "ReplaceGrantsRequest": {
        "description": "Запит на заміну набору ресурсних грантів ролі.",
        "properties": {
          "grants": {
            "description": "Новий набір; порожній прибирає доступ ролі повністю.",
            "items": {
              "$ref": "#/components/schemas/ResourceGrantDto"
            },
            "type": "array"
          }
        },
        "required": [
          "grants"
        ],
        "type": "object"
      },
...
      "ResourceGrantDto": {
        "description": "Ресурсний грант у вигляді, придатному для передавання.",
        "properties": {
          "isDeny": {
            "description": "Явна заборона; перекриває будь-який дозвіл (ФВ-6.6).",
            "type": "boolean"
          },
          "level": {
            "$ref": "#/components/schemas/GrantLevel",
            "description": "Рівень: `Read`…`Manage`."
          },
          "resourceId": {
            "description": "Ідентифікатор ресурсу.",
            "format": "int32",
            "type": "integer"
          },
          "resourceKind": {
            "$ref": "#/components/schemas/ResourceKind",
            "description": "Вид ресурсу: `Project`, `Sheet`, `Table`, `Column`."
          }
        },
        "required": [
          "resourceKind",
          "resourceId",
          "level",
          "isDeny"
        ],
        "type": "object"
      },
      "ResourceKind": {
        "description": "Тип ресурсу, на який видається грант.",
        "enum": [
          "Project",
          "Sheet",
          "Table",
          "Column"
        ]
      },
      "RoleIdResponse": {
        "description": "Створена роль.",
        "properties": {
          "roleId": {
            "description": "Ідентифікатор.",
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "roleId"
        ],
        "type": "object"
      },
      "RoleView": {
        "description": "Роль із її правами.",
        "properties": {
          "code": {
            "description": "Код.",
            "type": "string"
          },
          "dangerousPermissions": {
            "description": "Небезпечні права серед них — показуються окремо, бо їх видають поіменно.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "id": {
            "description": "Ідентифікатор.",
            "format": "int32",
            "type": "integer"
          },
          "isActive": {
            "description": "Чи діє.",
            "type": "boolean"
          },
          "isBuiltIn": {
            "description": "Вбудована роль із seed: видаленню не підлягає.",
            "type": "boolean"
          },
          "permissions": {
            "description": "Права ролі.",
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "id",
          "code",
          "isBuiltIn",
          "isActive",
          "permissions",
          "dangerousPermissions"
        ],
        "type": "object"
      },
...
      "SetAlertsRequest": {
        "description": "Запит на зміну отримання алертів.",
        "properties": {
          "receivesAlerts": {
            "description": "Чи отримує людина алерти про збої.",
            "type": "boolean"
          }
        },
        "required": [
          "receivesAlerts"
        ],
        "type": "object"
      },
...
      "SimulationSessionResponse": {
        "description": "Розпочатий сеанс симуляції.",
        "properties": {
          "readOnly": {
            "description": "Завжди `true`: будь-який запис під симуляцією відхиляється.",
            "type": "boolean"
          },
          "sessionId": {
            "description": "Сеанс; ним же він і завершується.",
            "format": "int64",
            "type": "integer"
          },
          "simulatedForUserId": {
            "description": "Чиїми правами дивимося.",
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "sessionId",
          "simulatedForUserId",
          "readOnly"
        ],
        "type": "object"
      },
...
      "StartSimulationRequest": {
        "description": "Запит на початок симуляції.",
        "properties": {
          "reason": {
            "description": "Причина; потрапляє в `aud.SimulationSession`.",
            "type": "string"
          },
          "subjectUserId": {
            "description": "Чиїми очима дивимося.",
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "subjectUserId",
          "reason"
        ],
        "type": "object"
      },
...
      "UserIdResponse": {
        "description": "Створений користувач.",
        "properties": {
          "userId": {
            "description": "Ідентифікатор.",
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "userId"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Ecr.Api | v1"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/auth/change-password": {
      "post": {
        "description": "Доступна навіть під `MustChangePassword`, коли решта запитів\r\nвідхиляється `ECR-PWD-0428` — інакше користувач не мав би способу\r\nвийти з цього стану.",
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "description": "Токен скасування.",
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Зміна власного пароля.",
        "tags": [
          "Security"
        ]
      }
    },
...
    "/api/v1/roles": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleView"
                  },
                  "type": "array"
                }
              },
              "text/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleView"
                  },
                  "type": "array"
                }
              },
              "text/plain": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleView"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Перелік ролей. Право `Security.ManageRoles`.",
        "tags": [
          "Security"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleIdResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleIdResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RoleIdResponse"
                }
              }
            },
            "description": "Created"
          }
        },
        "summary": "Створює роль. Право `Security.ManageRoles`.",
        "tags": [
          "Security"
        ]
      }
    },
    "/api/v1/roles/{id}/grants": {
      "get": {
        "description": "⛔ Без цих двох маршрутів система не показує даних НІКОМУ (`A7-22`):\r\nдоступ до проєкту, аркуша чи таблиці вимагає гранта, а створити грант\r\nне було чим. Права відповідають на питання «що людина вміє», гранти —\r\n«до чого саме»; без другої відповіді перша нічого не відкриває.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResourceGrantDto"
                  },
                  "type": "array"
                }
              },
              "text/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResourceGrantDto"
                  },
                  "type": "array"
                }
              },
              "text/plain": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResourceGrantDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Ресурсні гранти ролі. Право `Security.ManageRoles`.",
        "tags": [
          "Security"
        ]
      },
      "put": {
        "description": "⚠ Саме заміна набору, а не правка по одному: гранти — це відповідь на\r\nпитання «що покриває роль», і вона має бути видима одним поглядом.\r\nЧасткові правки лишають стан, у якому джерело доступу не відновлюється.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceGrantsRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceGrantsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceGrantsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "Замінює набір грантів ролі цілком. Право `Security.ManageRoles`.",
        "tags": [
          "Security"
        ]
      }
    },
    "/api/v1/security/simulation": {
      "delete": {
        "parameters": [
          {
            "description": "Сеанс.",
            "in": "query",
            "name": "sessionId",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "Завершує власний сеанс симуляції.",
        "tags": [
          "Security"
        ]
      },
      "post": {
        "description": "Сеанс пишеться в `aud.SimulationSession` <b>одразу</b> і не\r\nбатчиться: без запису «подивитися очима» стало б способом безслідно\r\nпереглянути чужі дані. Будь-який запис під симуляцією відхиляється\r\n`EditDenyReason.SimulationReadOnly` — навіть із правом\r\n`Manage` (ФВ-6.16a).",
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StartSimulationRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSimulationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationSessionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationSessionResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationSessionResponse"
                }
              }
            },
            "description": "Created"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Починає сеанс симуляції. Право `Security.Simulate`.",
        "tags": [
          "Security"
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfUserView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfUserView"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfUserView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Перелік користувачів. Право `Security.ManageUsers`.",
        "tags": [
          "Security"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserIdResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserIdResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserIdResponse"
                }
              }
            },
            "description": "Created"
          }
        },
        "summary": "Створює користувача. Право `Security.ManageUsers`.",
        "tags": [
          "Security"
        ]
      }
    },
    "/api/v1/users/{id}/alerts": {
      "put": {
        "description": "⚠ Адресати алертів — дані, а не конфігурація (`D-125`): перелік у\r\nзмінних оточення довелося б міняти розгортанням щоразу, коли хтось іде\r\nу відпустку.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetAlertsRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAlertsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAlertsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "Вмикає або вимикає отримання алертів. Право `Security.ManageUsers`.",
        "tags": [
          "Security"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
...
}